requirements.txt
faiss-cpu~=1.12.0
ollama~=0.5.3
openai~=1.107.2
pypdf~=6.0.0
python-dotenv~=1.1.1
requests~=2.32.5
sentence-transformers~=5.1.0
yfinance~=0.2.65